Parallel replication async deadlock kill
When a deadlock kill is detected inside the storage engine, the kill
is not done immediately, to avoid calling back into the storage engine
kill_query method with various lock subsystem mutexes held. Instead the
kill is queued and done later by a slave background thread.

This avoids the need for locking hacks in kill_query in both
InnoDB/XtraDB and TokuDB.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>