[MDEV-10570] Integrate existing Community Patch “Flashback” into MariaDB Server.
==== Description ====

Flashback can roll-back the instances/databases/tables to a snapshot.
It's implement on Server-Level by full image format binary logs, so it supports all engines.
Currently, it’s a feature inside mysqlbinlog tool (with --flashback arguments).

Because of the flashback binlog events will store in the memory, you should check if there is enough memory in your machine.

==== New Arguments ====

--flashback (-B)
It will let mysqlbinlog to work on FLASHBACK mode.

==== Example ====

I have a table "t" in database "test", we can compare the output with --flashback and without.
#client/mysqlbinlog /data/mysqldata_10.0/binlog/mysql-bin.000001 -vv -d test -T t --start-datetime="2013-03-27 14:54:00" > /tmp/1.sql
#client/mysqlbinlog /data/mysqldata_10.0/binlog/mysql-bin.000001 -vv -uroot -d test -T t --start-datetime="2013-03-27 14:54:00" -B > /tmp/2.sql

Then, importing the output flashback file (/tmp/2.log), it can flashback your database/table to the special time (--start-datetime).
And if you know the postion, --start-postion is also works, then mysqlbinlog will output the flashback logs that can flashback to --start-postion.

==== Implement ====

1. As we know, if binlog_format is ROW (binlog-row-image=FULL in 5.6 and later), all columns’ values are store in the row event, so we can get the data before mis-operation.

2. Just do following things:
  2.1 Change Event Type, INSERT->DELETE, DELETE->INSERT
  2.2 For Update_Event, swapping the SET part and WHERE part
  2.3 Applying those events from the last one to the first one which mis-operation happened.
  2.4 All the data will be recovered by inverse operations of mis-oprerations.